requirements.txt
uvicorn
mangum~=0.17.0
fastapi
pydub
openai-whisper
